# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://www.baidu.com"
        },
        "version": "2.0"
    },
    "host": "192.168.1.163:7001",
    "paths": {
...
        "/v1/fil_pool_ratio/create": {
            "post": {
                "description": "批量创建股东矿池比例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东矿池比例"
                ],
                "summary": "批量创建股东矿池比例",
                "parameters": [
                    {
                        "description": "创建股东矿池比例 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.CreateFilPoolRatio"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool_ratio/detail": {
            "get": {
                "description": "根据矿池id获取股东矿池列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东矿池比例"
                ],
                "summary": "根据矿池id获取股东矿池列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool_ratio/list": {
            "get": {
                "description": "获取股东矿池比例数据列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股东矿池比例"
                ],
                "summary": "获取股东矿池比例数据列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id 可选参数",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取当页条数",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateFilPoolRatio": {
            "type": "object",
            "required": [
                "fil_pool_id",
                "list"
            ],
            "properties": {
                "fil_pool_id": {
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.ShareholerShareList"
                    }
                }
            }
        },
...
        "request.ShareholerShareList": {
            "type": "object",
            "required": [
                "proportion_of_shares",
                "shareholder_id"
            ],
            "properties": {
                "proportion_of_shares": {
                    "type": "integer"
                },
                "shareholder_id": {
                    "type": "string"
                }
            }
        },
...
        "response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}